+++ ocd-960.sql
@@ -33,3 +33,5 @@ $$ LANGUAGE plpgsql;
 select openchpl.deleteDuplicateCertifiedProducts();
 
 DROP FUNCTION openchpl.deleteDuplicateCertifiedProducts();
+
+