+++ dev/openchpl_data-model.sql
@@ -2146,6 +2146,130 @@ CREATE TABLE openchpl.api_key_activity
 
 ALTER TABLE openchpl.api_key_activity ADD CONSTRAINT api_key_fk FOREIGN KEY (api_key_id) REFERENCES openchpl.api_key (api_key_id);
 
+CREATE TABLE openchpl.surveillance_type (
+	id bigserial not null,
+	name varchar(50) not null,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_type_pk PRIMARY KEY (id),
+	CONSTRAINT surveillance_type_name_key UNIQUE (name)
+);
+
+CREATE TABLE openchpl.surveillance_requirement_type (
+	id bigserial not null,
+	name varchar(100) not null,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_requirement_type_pk PRIMARY KEY (id),
+	CONSTRAINT surveillance_requirement_type_name_key UNIQUE (name)
+);
+
+CREATE TABLE openchpl.surveillance_result (
+	id bigserial not null,
+	name varchar(100) not null,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_result_pk PRIMARY KEY (id),
+	CONSTRAINT surveillance_result_name_key UNIQUE (name)
+);
+
+CREATE TABLE openchpl.nonconformity_status (
+	id bigserial not null,
+	name varchar(50) not null,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT nonconformity_status_pk PRIMARY KEY (id),
+	CONSTRAINT nonconformity_status_name_key UNIQUE (name)
+);
+
+CREATE TABLE openchpl.surveillance (
+	id bigserial not null,
+	certified_product_id bigint not null,
+	start_date date not null,
+	end_date date,
+	type_id bigint not null,
+	randomized_sites_used integer, -- required if type is Randomized
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_pk PRIMARY KEY (id),
+	CONSTRAINT certified_product_fk FOREIGN KEY (type_id) 
+		REFERENCES openchpl.certified_product (certified_product_id) 
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,	
+	CONSTRAINT type_fk FOREIGN KEY (type_id) 
+		REFERENCES openchpl.surveillance_type (id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE		
+);
+
+CREATE TABLE openchpl.surveillance_requirement (
+	id bigserial not null,
+	surveillance_id bigint not null,
+	type_id bigint not null,
+	-- either criteria or requirement text is required
+	certification_criterion_id bigint,
+	requirement varchar(1024),
+	result_id bigint, -- required if parent surveillance has an end date
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_requirement_id PRIMARY KEY (id),
+	CONSTRAINT surveillance_fk FOREIGN KEY (surveillance_id) 
+		REFERENCES openchpl.surveillance (id) 
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,
+	CONSTRAINT type_fk FOREIGN KEY (type_id) 
+		REFERENCES openchpl.surveillance_requirement_type (id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,
+	CONSTRAINT certification_criterion_fk FOREIGN KEY (certification_criterion_id) 
+		REFERENCES openchpl.certification_criterion (certification_criterion_id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,		
+	CONSTRAINT result_fk FOREIGN KEY (result_id) 
+		REFERENCES openchpl.surveillance_result (id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE		
+);
+
+CREATE TABLE openchpl.surveillance_nonconformity (
+	id bigserial not null,
+	surveillance_requirement_id bigint not null,
+	-- either criteria or type is required
+	certification_criterion_id bigint,
+	nonconformtiy_type varchar(1024), 
+	nonconformity_status_id bigint not null,
+	date_of_determination date not null,
+	corrective_action_plan_approval_date date,
+	corrective_action_start_date date,
+	corrective_action_must_complete_date date,
+	corrective_action_end_date date,
+	summary varchar(2048) not null, -- required if nonconformity type+certification_criterion is not null
+	findings text not null, -- required if nonconformity type+certification_criterion is not null
+	sites_passed integer, -- only pertintent to Randomized surveillance
+	total_sites integer,
+	developer_explanation text,
+	resolution text, -- required if status is Closed
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT surveillance_nonconformity_pk PRIMARY KEY (id),
+	CONSTRAINT surveillance_requirement_fk FOREIGN KEY (surveillance_requirement_id) 
+		REFERENCES openchpl.surveillance_requirement (id) 
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,
+	CONSTRAINT certification_criterion_fk FOREIGN KEY (certification_criterion_id) 
+		REFERENCES openchpl.certification_criterion (certification_criterion_id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,
+	CONSTRAINT nonconformity_status_fk FOREIGN KEY (nonconformity_status_id) 
+		REFERENCES openchpl.nonconformity_status (id) 
+		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE	
+);
 
 -- Table: openchpl.ehr_certification_id