+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ _Date TBD_
 ### Changes
 * Added script to remove three specific duplicate certified products
 * Added product owner history relationship
+* Add surveillance tables to the database 
 
 ---